pyproject.toml
[project]
name = "splax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jonah Turner", email = "jonahturner2000@hotmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.1.0",
    "flax>=0.10.1",
    "hydra-core>=1.3.2",
    "jax[cuda12]>=0.4.35",
    "jaxtyping>=0.2.34",
    "optax>=0.2.3",
    "orbax>=0.1.9",
    "torch>=2.5.1",
    "tqdm>=4.66.6",
    "transformers>=4.46.1",
    "wandb>=0.18.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
